requirements.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
flask
requests
pyyaml

## -e .